some login

+++ src/categories/category/category.component.ts
@@ -2,6 +2,7 @@ import {Component, OnInit} from '@angular/core';
 import {Category} from '../../model/category';
 import {Router} from '@angular/router';
 import {CategoriesService} from '../../services/categories.service';
+import {ConfirmationService} from 'primeng/api';
 
 @Component({
   selector: 'app-category',
@@ -12,7 +13,8 @@ export class CategoryComponent implements OnInit {
 
   constructor(
     private router: Router,
-    private service: CategoriesService) {
+    private service: CategoriesService,
+    private confirmationService: ConfirmationService) {
   }
 
   public categories: Category[] = [];
@@ -25,8 +27,15 @@ export class CategoryComponent implements OnInit {
     this.service.getAll().subscribe(value => this.categories = value);
   }
 
-  delete(id: number): void {
-    this.service.delete(id).subscribe(value => this.refreshList());
+
+  delete(id: number): boolean {
+    this.confirmationService.confirm({
+      message: 'Are you sure you want to delete this item?',
+      accept: () => {
+        this.service.delete(id).subscribe(value => this.refreshList());
+      }
+    });
+    return false;
   }
 
   add(): void {
